% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{decrypt_conn}
\alias{decrypt_conn}
\alias{decrypt_file}
\title{Decrypt a connection or file}
\usage{
decrypt_conn(encryptedConn, plainConn, key = "pass", method = "symmetric")
}
\arguments{
\item{encryptedConn}{A text connection or a file where the encrypted string
was written}

\item{plainConn}{A raw connection or a file where decrypted content will be
written to}

\item{key}{A string without embbeded NULL. Default is 'pass'.}

\item{method}{Currently, a stub. It should be 'symmetric'(default)}
}
\value{
An invisible TRUE
}
\description{
\code{decrypt_conn}(or \code{decrypt_file}) decrypts a text
  connection or a file to a raw connection or a file. \code{decrypt_conn}(or
  \code{decrypt_file}) decrypts a text connection or a file (encrypted by
  \code{encrypt_conn} or \code{encrypt_file})
}
