% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{decrypt_object}
\alias{decrypt_object}
\alias{unserialize2}
\title{Decrypt an encrypted R object}
\usage{
decrypt_object(encryptedObject, key = "pass", method = "symmetric")
}
\arguments{
\item{encryptedObject}{A raw object or a string to be decrypted}

\item{key}{A string without embbeded NULL. Default is 'pass'.}

\item{method}{Currently, a stub. It should be 'symmetric'(default)}
}
\value{
A R object
}
\description{
\code{encrypt_object}(or \code{serialize2}) encrypts a R object
  as a raw object or a string and \code{decrypt_object}(or
  \code{unserialize2}) decrypts a raw object or a string(encrypted by
  \code{encrypt_object})
}
