% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{encrypt_conn}
\alias{encrypt_conn}
\alias{encrypt_file}
\title{Encrypt a connection or file}
\usage{
encrypt_conn(plainConn, encryptedConn, key = "pass", method = "symmetric")
}
\arguments{
\item{plainConn}{A raw connection or a file to be encrypted}

\item{encryptedConn}{A text connection or a file where the encrypted string
is to be written}

\item{key}{A string without embbeded NULL. Default is 'pass'.}

\item{method}{Currently, a stub. It should be 'symmetric'(default)}
}
\value{
An invisible TRUE
}
\description{
\code{encrypt_conn}(or \code{encrypt_file}) encrypts a raw
  connection to a text connection or a file. \code{decrypt_conn}(or
  \code{decrypt_file}) decrypts a text connection or a file (encrypted by
  \code{encrypt_conn} or \code{encrypt_file})
}
