% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{encrypt_object}
\alias{encrypt_object}
\alias{serialize2}
\title{Encrypt a R object}
\usage{
encrypt_object(object, ascii = FALSE, key = "pass", method = "symmetric")
}
\arguments{
\item{object}{A R object to be encrypted}

\item{ascii}{\code{TRUE} if the R object is to be encrypted as a string.
Default is \code{FALSE}}

\item{key}{A string without embbeded NULL. Default is 'pass'.}

\item{method}{Currently, a stub. It should be 'symmetric'(default)}
}
\value{
A raw object if \code{ascii} is \code{FALSE}. A string if
  \code{ascii} is \code{TRUE}.
}
\description{
\code{encrypt_object}(or \code{serialize2}) encrypts a R object
  as a raw object or a string and \code{decrypt_object}(or
  \code{unserialize2}) decrypts a raw object or a string(encrypted by
  \code{encrypt_object})
}
