% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize SAFER Database}
\usage{
init(
  df,
  db_name = "safer.sqlite",
  table_name = "safer_table",
  append = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{df}{Dataframe to be stored in the database.}

\item{db_name}{Name of the SQLite database file. Default is "safer.sqlite".}

\item{table_name}{Name of the table within the database. Default is "safer_table".}

\item{append}{Logical; if TRUE, data will be appended to the existing table. Default is FALSE.}

\item{overwrite}{Logical; if TRUE, existing table will be overwritten. Default is FALSE.}
}
\value{
None
}
\description{
This function initializes a SQLite database for the SAFER process.
It connects to a database, modifies the input dataframe by adding an index and a label column,
and then writes the data into a specified table.
}
\examples{
\dontrun{
init(df = your_dataframe)
}
}
