% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypair.R
\name{keypair}
\alias{keypair}
\title{Generate a public key and private key pair}
\usage{
keypair(seed = NULL)
}
\arguments{
\item{seed}{A raw object. If NULL, a randon seed will be chosen.}
}
\value{
A list with:
\itemize{
\item public_key: A raw object
\item private_key: A raw object
\item seed: A raw object
}
}
\description{
Using sodium's 'keygen' and 'pubkey' based on curve25519
}
\examples{
temp <- keypair()
str(temp)
}
