% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_crime.R
\name{mean_crime}
\alias{mean_crime}
\title{Title Calculates the mean crime in a specific geographic area}
\usage{
mean_crime(station)
}
\arguments{
\item{station}{A valid station in the crime data frame, string}
}
\value{
The average of all the crime rates in a specific Station and Category
}
\description{
Title Calculates the mean crime in a specific geographic area
}
\examples{

mean_crime("Cape Town Central")
1


}
