% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_crime.R
\name{plotting_crime}
\alias{plotting_crime}
\title{Plot Types of Crime per Province of a Given Year}
\usage{
plotting_crime(year)
}
\arguments{
\item{year}{A specified year (found in the Year variable in the crime dataset)}
}
\value{
A bargraph of the types of crime per province of a given year
}
\description{
Plot Types of Crime per Province of a Given Year
}
\examples{
plotting_crime(2005)
}
