% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process Rows in SAFER Database}
\usage{
process(db_name = "safer.sqlite", table_name = "safer_table", .f)
}
\arguments{
\item{db_name}{Name of the SQLite database file. Default is "safer.sqlite".}

\item{table_name}{Name of the table within the database. Default is "safer_table".}

\item{.f}{A user-defined function to process the text column in each row.}
}
\value{
None
}
\description{
This function processes rows with NA labels in a specified table within a SQLite database.
It repeatedly selects a row with NA label, applies a user-defined function to it, and updates the row in the database.
}
\examples{
\dontrun{
process(con = NULL, .f = your_processing_function)
}
}
