% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.R
\name{safe}
\alias{safe}
\title{Safe piping}
\usage{
safe(result)
}
\arguments{
\item{result}{object, thing to send into the safe pipe}

\item{error}{error, thing to send into the safe pipe}
}
\value{
safe object
}
\description{
This S3 object is inspired by \code{\link[purrr:safely]{purrr::safely()}}.
}
\details{
It is a list with elements:
\itemize{
\item \code{result}
\item \code{error}
}

One of these, at most, will be not \code{NULL}
}
