% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATion.R
\name{STATion}
\alias{STATion}
\title{Title Crimes Most Frequently Committed in a Given Station, Year}
\usage{
STATion(station, year)
}
\arguments{
\item{station}{A specified South African Police Station (found in the Station variable in the crime dataset)}

\item{year}{A specified year (found in the Year variable in the crime dataset)}
}
\value{
A barplot of the top three most frequently committed crimes in a specified year and station.
}
\description{
Title Crimes Most Frequently Committed in a Given Station, Year
}
\examples{
max_crime("Cape Town Central", 2005)
max_crime("Park Road", 2010)
}
