% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_line.R
\name{trend_line}
\alias{trend_line}
\title{Title Create a trend line of all the crime rates in a specific geographic area}
\usage{
trend_line(station)
}
\arguments{
\item{station}{A valid station in the crime data frame, string}
}
\value{
A visualization of all the crime rates in a specific Station and Category
}
\description{
Title Create a trend line of all the crime rates in a specific geographic area
}
